requirements.txt
# =============================================================================
# ScholarRAG 依赖文件
# 安装命令: pip install -r requirements.txt
# =============================================================================

# =============================================================================
# Web 框架和服务器
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# =============================================================================
# 数据库
# =============================================================================
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# =============================================================================
# 向量数据库
# =============================================================================
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0

# =============================================================================
# LangChain / LangGraph 生态
# =============================================================================
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langgraph>=0.0.50

# =============================================================================
# 数据验证和配置
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# =============================================================================
# 工具和实用库
# =============================================================================
tiktoken>=0.5.0
httpx>=0.25.0
aiofiles>=23.2.0
pillow>=10.0.0
numpy>=1.24.0
prometheus-client>=0.20.0
tenacity>=9.0.0

# =============================================================================
# 开发依赖 (测试和代码质量)
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
